perbaikan navbar admin, pembuatan migrate postsblog, perbaikan tampilan produk admin, crud blog kategori, crud blog

+++ app/Http/Controllers/Admin/PostCategoriesController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers\Admin;
 use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
 use App\Models\PostCategories;
+use App\Models\Posts;
 class PostCategoriesController extends Controller
 {
     public function index()
@@ -39,11 +40,22 @@ class PostCategoriesController extends Controller
         return view('categories.show', compact('category'));
     }
 
-    public function edit($id)
-    {
-        $category = PostCategories::find($id);
-        return view('categories.edit', compact('category'));
+    // public function edit($id)
+    // {
+    //     $categories = PostCategories::find($id);
+    //     //dd($categories);
+    //     return view('admin.blog-categories.edit-blog-kategori', compact('categories'));
 
+    // }
+
+    public function edit($id){
+        $data = [
+            "edit" => PostCategories::where("id", $id)->first(),
+            "data" => PostCategories::where("id", "!=" , $id)->get()
+
+        ];
+
+        return view("admin.blog-categories.edit-blog-kategori", $data);
     }
 
     public function hapus(Request $request)
@@ -55,15 +67,43 @@ class PostCategoriesController extends Controller
             return redirect()->route('posts.kategori')->with('gagal', 'Kategori tidak ditemukan');
         }
 
-        // cek apakah kategori terkait sudah ada di tabel produk
-        // $productCount = Produk::where('categories_id', $id)->count();
+        //cek apakah kategori terkait sudah ada di tabel produk
+        $postsCount = Posts::where('category_id', $id)->count();
 
-        // if ($productCount > 0) {
-        //     return redirect()->route('posts.kategori')->with('gagal', 'Kategori terkait masih digunakan di ....');
-        // }
+        if ($postsCount > 0) {
+            return redirect()->route('posts.kategori')->with('gagal', 'Kategori terkait masih digunakan di ....');
+        }
 
         $kategori->delete();
 
         return redirect()->route('posts.kategori')->with('sukses', 'Kategori berhasil dihapus');
+    }
+
+    public function update(Request $request)
+    {
+        $message = [
+            "required" => "Kolom :attribute wajib diisi",
+            'min' => "kolom :attribute minimal harus :min karakter",
+            'max' => "kolom :attribute maximal harus :max karakter"
+
+        ];
+
+        $this->validate($request, [
+            "name" => "required|min:4"
+        ], $message);
+
+        $cek_double = PostCategories::where(["name" => $request->name])->count();
+
+        if ($cek_double > 0) {
+            return redirect()->back()->with("gagal", "Tidak Boleh Duplikasi Data");;
+        }
+
+        PostCategories::where("id", $request->id)->update([
+            "name" => $request->name
+        ]);
+
+       // return response()->json(['messege' => 'request success'],200);
+       // return response()->json(['messege' => 'request success'],200);
+       return redirect()->route('posts.kategori')->with('sukses','data berhasil di ubah');
     }
 }
